NFT-999: Fix write button not working after approvals
