+++ components/Controllers/OpenVault/VaultDebtPicker/VaultDebtPicker.tsx
@@ -559,8 +559,6 @@ export function VaultDebtPicker({
               collateralContractAddress={collateralContractAddress}
               depositNFTs={depositNFTs}
               withdrawNFTs={withdrawNFTs}
-              vaultCollateralCount={vault?.collateralCount || 0}
-              vaultHasDebt={vaultHasDebt}
               errorMessage={errorMessage}
               setErrorMessage={setErrorMessage}
               refresh={refresh}
