+++ components/Controllers/OpenVault/VaultWriteButton.tsx
@@ -2,10 +2,12 @@ import { TransactionButton } from 'components/Button';
 import { ethers } from 'ethers';
 import { useAsyncValue } from 'hooks/useAsyncValue';
 import { useController } from 'hooks/useController';
+import { useOracleSynced } from 'hooks/useOracleSynced';
 import { usePaprBalance } from 'hooks/usePaprBalance';
 import { useSignerOrProvider } from 'hooks/useSignerOrProvider';
 import { VaultWriteType } from 'hooks/useVaultWrite/helpers';
 import { useVaultWrite } from 'hooks/useVaultWrite/useVaultWrite';
+import { OraclePriceType } from 'lib/oracle/reservoir';
 import { useEffect, useMemo, useState } from 'react';
 import { ERC20__factory } from 'types/generated/abis';
 import { useAccount } from 'wagmi';
@@ -18,10 +20,8 @@ type VaultWriteButtonProps = {
   collateralContractAddress: string;
   depositNFTs: string[];
   withdrawNFTs: string[];
-  vaultCollateralCount: number;
   amount: ethers.BigNumber;
   quote: ethers.BigNumber | null;
-  vaultHasDebt: boolean;
   errorMessage: string;
   setErrorMessage: (message: string) => void;
   refresh: () => void;
@@ -32,10 +32,8 @@ export function VaultWriteButton({
   collateralContractAddress,
   depositNFTs,
   withdrawNFTs,
-  vaultCollateralCount,
   amount,
   quote,
-  vaultHasDebt,
   errorMessage,
   setErrorMessage,
   refresh,
@@ -43,20 +41,87 @@ export function VaultWriteButton({
   const { address } = useAccount();
   const signerOrProvider = useSignerOrProvider();
   const paprController = useController();
+  const oracleInfoSynced = useOracleSynced(
+    collateralContractAddress,
+    OraclePriceType.lower,
+  );
 
-  const { data, write, error, usingSafeTransferFrom, oracleInfoSynced } =
-    useVaultWrite(
-      writeType,
-      collateralContractAddress,
-      depositNFTs,
-      withdrawNFTs,
-      amount,
-      quote,
-      refresh,
-    );
   const [collateralApproved, setCollateralApproved] = useState<boolean>(false);
   const [underlyingApproved, setUnderlyingApproved] = useState<boolean>(false);
   const [debtTokenApproved, setDebtTokenApproved] = useState<boolean>(false);
+
+  const usingSafeTransferFrom = useMemo(() => {
+    if (
+      depositNFTs.length === 1 &&
+      withdrawNFTs.length === 0 &&
+      (writeType === VaultWriteType.Borrow ||
+        writeType === VaultWriteType.BorrowWithSwap)
+    )
+      return true;
+    else return false;
+  }, [depositNFTs.length, withdrawNFTs.length, writeType]);
+
+  const disabled = useMemo(() => {
+    if (!oracleInfoSynced) return true;
+    switch (writeType) {
+      case VaultWriteType.Borrow:
+        return (
+          !usingSafeTransferFrom &&
+          !collateralApproved &&
+          depositNFTs.length > 0
+        );
+      case VaultWriteType.BorrowWithSwap:
+        return (
+          (!usingSafeTransferFrom &&
+            !collateralApproved &&
+            depositNFTs.length > 0) ||
+          !quote
+        );
+      case VaultWriteType.Repay:
+        return !!errorMessage || !debtTokenApproved;
+      case VaultWriteType.RepayWithSwap:
+        return !!errorMessage || !underlyingApproved || !quote;
+    }
+  }, [
+    oracleInfoSynced,
+    depositNFTs,
+    writeType,
+    usingSafeTransferFrom,
+    underlyingApproved,
+    collateralApproved,
+    debtTokenApproved,
+    errorMessage,
+    quote,
+  ]);
+
+  const buttonText = useMemo(() => {
+    if (!oracleInfoSynced) return 'Waiting for oracle...';
+    let actions: string[] = [];
+    if (depositNFTs.length > 0) actions = [...actions, 'Deposit'];
+    if (withdrawNFTs.length > 0) actions = [...actions, 'Withdraw'];
+    actions = [
+      ...actions,
+      writeType === VaultWriteType.Repay ||
+      writeType === VaultWriteType.RepayWithSwap
+        ? 'Repay'
+        : 'Borrow',
+    ];
+    if (actions.length > 2) return 'Update Loan';
+    else if (actions.length > 1) return actions.join(' & ');
+    else return actions[0];
+  }, [oracleInfoSynced, depositNFTs.length, withdrawNFTs.length, writeType]);
+
+  const { data, write, error } = useVaultWrite(
+    writeType,
+    collateralContractAddress,
+    depositNFTs,
+    withdrawNFTs,
+    amount,
+    quote,
+    usingSafeTransferFrom,
+    disabled,
+    refresh,
+  );
 
   const underlyingBalance = useAsyncValue(async () => {
     if (!address) return null;
@@ -109,56 +174,6 @@ export function VaultWriteButton({
     amount,
     quote,
   ]);
-
-  const disabled = useMemo(() => {
-    if (!oracleInfoSynced) return true;
-    switch (writeType) {
-      case VaultWriteType.Borrow:
-        return (
-          !usingSafeTransferFrom &&
-          !collateralApproved &&
-          depositNFTs.length > 0
-        );
-      case VaultWriteType.BorrowWithSwap:
-        return (
-          (!usingSafeTransferFrom &&
-            !collateralApproved &&
-            depositNFTs.length > 0) ||
-          !quote
-        );
-      case VaultWriteType.Repay:
-        return !!errorMessage || !debtTokenApproved;
-      case VaultWriteType.RepayWithSwap:
-        return !!errorMessage || !underlyingApproved || !quote;
-    }
-  }, [
-    oracleInfoSynced,
-    depositNFTs,
-    writeType,
-    usingSafeTransferFrom,
-    underlyingApproved,
-    collateralApproved,
-    debtTokenApproved,
-    errorMessage,
-    quote,
-  ]);
-
-  const buttonText = useMemo(() => {
-    if (!oracleInfoSynced) return 'Waiting for oracle...';
-    let actions: string[] = [];
-    if (depositNFTs.length > 0) actions = [...actions, 'Deposit'];
-    if (withdrawNFTs.length > 0) actions = [...actions, 'Withdraw'];
-    actions = [
-      ...actions,
-      writeType === VaultWriteType.Repay ||
-      writeType === VaultWriteType.RepayWithSwap
-        ? 'Repay'
-        : 'Borrow',
-    ];
-    if (actions.length > 2) return 'Update Loan';
-    else if (actions.length > 1) return actions.join(' & ');
-    else return actions[0];
-  }, [oracleInfoSynced, depositNFTs.length, withdrawNFTs.length, writeType]);
 
   return (
     <>
